Class_JS/class07-1/drf_backend_file/drf_backend_file/requirements.txt
asgiref==3.6.0; python_version >= "3.8"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
django-cors-headers==3.13.0; python_version >= "3.7"
django==4.1.5; python_version >= "3.8"
djangorestframework-simplejwt==5.2.2; python_version >= "3.7"
djangorestframework==3.14.0; python_version >= "3.6"
pyjwt==2.6.0; python_version >= "3.7"
pytz==2022.7.1; python_version >= "3.7"
sqlparse==0.4.3; python_version >= "3.8"
tzdata==2022.7; sys_platform == "win32" and python_version >= "3.8"
